+++ src/types/trades/failed.rs
@@ -1,10 +1,8 @@
-use crate::traits::AsDataFrame;
 use crate::types::reason_code::ReasonCode;
 use crate::types::signals::Side;
 use crate::types::trades::future::FutureTrade;
 use crate::types::trades::{calc_cost, Trade};
 use chrono::NaiveDateTime;
-use polars::frame::DataFrame;
 use polars::prelude::{NamedFrom, Series};
 
 /// Represents a trade that has been rejected by the market or otherwise failed
@@ -45,20 +43,6 @@ impl FailedTrade {
             cost: trade.get_cost(),
             point: trade.get_point().clone(),
         }
-    }
-}
-
-impl AsDataFrame for FailedTrade {
-    fn as_dataframe(&self) -> DataFrame {
-        DataFrame::new(vec![
-            Series::new("side", vec![self.side as i32]),
-            Series::new("price", vec![self.price]),
-            Series::new("quantity", vec![self.quantity]),
-            Series::new("cost", vec![self.cost]),
-            Series::new("reason", vec![self.reason as i32]),
-            Series::new("point", vec![self.point]),
-        ])
-        .unwrap()
     }
 }
 
@@ -131,66 +115,5 @@ mod test {
         assert_eq!(failed_trade.quantity, quantity);
         assert_eq!(failed_trade.cost, cost);
         assert_eq!(failed_trade.point, point);
-    }
-
-    /// Test the `as_dataframe` method for `FailedTrade`
-    #[test]
-    fn test_as_dataframe() {
-        let reason = ReasonCode::Unknown;
-        let side = Side::Buy;
-        let price = 1.0;
-        let quantity = 2.0;
-        let cost = 3.0;
-        let point = NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap();
-
-        let trade = FailedTrade {
-            reason,
-            side,
-            price,
-            quantity,
-            cost,
-            point,
-        };
-
-        let df = trade.as_dataframe();
-        assert_eq!(df.shape(), (1, 6));
-        assert_eq!(
-            df.get_column_names(),
-            &["side", "price", "quantity", "cost", "reason", "point"]
-        );
-        assert_eq!(
-            df.column("side").unwrap().i32().unwrap().get(0).unwrap(),
-            side as i32
-        );
-        assert_eq!(
-            df.column("price").unwrap().f64().unwrap().get(0).unwrap(),
-            price
-        );
-        assert_eq!(
-            df.column("quantity")
-                .unwrap()
-                .f64()
-                .unwrap()
-                .get(0)
-                .unwrap(),
-            quantity
-        );
-        assert_eq!(
-            df.column("cost").unwrap().f64().unwrap().get(0).unwrap(),
-            cost
-        );
-        assert_eq!(
-            df.column("reason").unwrap().i32().unwrap().get(0).unwrap(),
-            reason as i32
-        );
-        assert_eq!(
-            df.column("point")
-                .unwrap()
-                .datetime()
-                .unwrap()
-                .last()
-                .unwrap(),
-            point.timestamp_millis()
-        );
     }
 }