+++ src/types/trades/future.rs
@@ -41,19 +41,6 @@ impl FutureTrade {
     }
 }
 
-impl AsDataFrame for FutureTrade {
-    fn as_dataframe(&self) -> DataFrame {
-        DataFrame::new(vec![
-            Series::new("side", vec![self.side as i32]),
-            Series::new("price", vec![self.price]),
-            Series::new("quantity", vec![self.quantity]),
-            Series::new("cost", vec![self.cost]),
-            Series::new("point", vec![self.point]),
-        ])
-        .unwrap()
-    }
-}
-
 impl Trade for FutureTrade {
     fn get_side(&self) -> Side {
         self.side
@@ -78,7 +65,6 @@ impl Trade for FutureTrade {
 
 #[cfg(test)]
 mod tests {
-    use crate::traits::AsDataFrame;
     use crate::types::signals::Side;
     use crate::types::trades::future::FutureTrade;
     use crate::types::trades::Trade;
@@ -98,52 +84,5 @@ mod tests {
         assert_eq!(trade.get_quantity(), quantity);
         assert_eq!(trade.get_cost(), price * quantity);
         assert_eq!(trade.get_point(), &point);
-    }
-
-    #[test]
-    fn test_as_dataframe() {
-        let side = Side::Buy;
-        let price = 1.0;
-        let quantity = 2.0;
-        let point = NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap();
-
-        let trade = FutureTrade::new(side, price, quantity, point);
-        let df = trade.as_dataframe();
-
-        assert_eq!(df.shape(), (1, 5));
-        assert_eq!(
-            df.get_column_names(),
-            &["side", "price", "quantity", "cost", "point"]
-        );
-        assert_eq!(
-            df.column("side").unwrap().i32().unwrap().get(0).unwrap(),
-            side as i32
-        );
-        assert_eq!(
-            df.column("price").unwrap().f64().unwrap().get(0).unwrap(),
-            price
-        );
-        assert_eq!(
-            df.column("quantity")
-                .unwrap()
-                .f64()
-                .unwrap()
-                .get(0)
-                .unwrap(),
-            quantity
-        );
-        assert_eq!(
-            df.column("cost").unwrap().f64().unwrap().get(0).unwrap(),
-            trade.cost
-        );
-        assert_eq!(
-            df.column("point")
-                .unwrap()
-                .datetime()
-                .unwrap()
-                .get(0)
-                .unwrap(),
-            point.timestamp_millis()
-        );
     }
 }